feat(frontend): Refator
