+++ frontend/src/lib/components/apps/components/table/AppTable.svelte
@@ -6,49 +6,53 @@
 	import type { AppInput } from '../../inputType'
 	import RunnableWrapper from '../helpers/RunnableWrapper.svelte'
 	import { writable } from 'svelte/store'
-	import {
-		createSvelteTable,
-		flexRender,
-		getCoreRowModel,
-		type TableOptions,
-		getPaginationRowModel
-	} from '@tanstack/svelte-table'
+	import { createSvelteTable, flexRender, type TableOptions } from '@tanstack/svelte-table'
 	import AppButton from '../buttons/AppButton.svelte'
 	import { classNames } from '$lib/utils'
-	import Button from '$lib/components/common/button/Button.svelte'
-	import { faDownload, faRefresh } from '@fortawesome/free-solid-svg-icons'
 	import DebouncedInput from '../helpers/DebouncedInput.svelte'
+	import AppTableFooter from './AppTableFooter.svelte'
+	import RefreshButton from '../helpers/RefreshButton.svelte'
+	import { tableOptions } from './tableOptions'
 
 	export let id: string
 	export let componentInput: AppInput | undefined
 	export let configuration: Record<string, AppInput>
 	export let actionButtons: (BaseAppComponent & ButtonComponent)[]
 
+	export const staticOutputs: string[] = ['selectedRow', 'loading', 'result']
+
 	type T = Record<string, any>
 
 	$: result = [] as Array<Record<string, any>>
 
 	let search: 'Frontend' | 'Backend' | 'Disabled' = 'Disabled'
-	let pagination: boolean | undefined = undefined
-
-	let page = 1
 	let searchValue = ''
 
-	const tableOptions = {
-		getCoreRowModel: getCoreRowModel(),
-		getPaginationRowModel: getPaginationRowModel(),
-		initialState: {
-			pagination: {
-				pageSize: 10
-			}
-		}
-	}
+	let pagination: boolean | undefined = undefined
+	let page = 1
 
 	const options = writable<TableOptions<T>>({
 		data: result,
 		columns: [],
 		...tableOptions
 	})
+
+	const table = createSvelteTable(options)
+
+	const { worldStore, staticOutputs: staticOutputsStore } =
+		getContext<AppEditorContext>('AppEditorContext')
+
+	let selectedRowIndex = -1
+
+	function toggleRow(row: Record<string, any>, rowIndex: number) {
+		if (selectedRowIndex === rowIndex) {
+			selectedRowIndex = -1
+			outputs.selectedRow.set(null)
+		} else {
+			selectedRowIndex = rowIndex
+			outputs?.selectedRow.set(row.original)
+		}
+	}
 
 	function setOptions(filteredResult: Array<Record<string, any>>) {
 		const headers = Array.from(new Set(result.flatMap((row) => Object.keys(row))))
@@ -65,38 +69,6 @@
 		}
 	}
 
-	$: filteredResult && setOptions(filteredResult)
-
-	const table = createSvelteTable(options)
-
-	const {
-		worldStore,
-		staticOutputs: staticOutputsStore,
-		runnableComponents
-	} = getContext<AppEditorContext>('AppEditorContext')
-
-	export const staticOutputs: string[] = ['selectedRow', 'loading', 'result']
-
-	$: outputs = $worldStore?.outputsById[id] as {
-		selectedRow: Output<any>
-	}
-
-	let selectedRowIndex = -1
-
-	function toggleRow(row: Record<string, any>, rowIndex: number) {
-		if (selectedRowIndex === rowIndex) {
-			selectedRowIndex = -1
-			outputs.selectedRow.set(null)
-		} else {
-			selectedRowIndex = rowIndex
-			outputs?.selectedRow.set(row.original)
-		}
-	}
-
-	$: extraQueryParams = search === 'Backend' ? { search: searchValue, page } : { page, search: '' }
-
-	$: console.log({ extraQueryParams })
-
 	function searchInResult(result: Array<Record<string, any>>, searchValue: string) {
 		if (searchValue === '') {
 			return result
@@ -108,14 +80,12 @@
 
 	let filteredResult: Array<Record<string, any>> = []
 
+	$: filteredResult && setOptions(filteredResult)
+	$: extraQueryParams = search === 'Backend' ? { search: searchValue, page } : { page, search: '' }
 	$: search === 'Frontend' && (filteredResult = searchInResult(result, searchValue))
 	$: (search === 'Backend' || search === 'Disabled') && (filteredResult = result)
-
-	let loading = false
-	async function refresh() {
-		loading = true
-		await $runnableComponents[id]?.()
-		loading = false
+	$: outputs = $worldStore?.outputsById[id] as {
+		selectedRow: Output<any>
 	}
 </script>
 
@@ -126,18 +96,10 @@
 	<div class="border border-gray-300 shadow-sm divide-y divide-gray-300  flex flex-col h-full">
 		<div class="py-2 px-4">
 			<div class="flex justify-between items-center">
-				<Button
-					iconOnly
-					startIcon={{ icon: faRefresh, classes: loading ? 'animate-spin' : '' }}
-					color="dark"
-					size="xs"
-					on:click={refresh}
-				/>
+				<RefreshButton componentId={id} />
 				{#if search !== 'Disabled'}
 					<div>
-						<div>
-							<DebouncedInput placeholder="Search..." bind:value={searchValue} />
-						</div>
+						<DebouncedInput placeholder="Search..." bind:value={searchValue} />
 					</div>
 				{/if}
 			</div>
@@ -202,52 +164,7 @@
 				</tbody>
 			</table>
 		</div>
-		<div class="px-4 py-2 text-xs flex flex-row gap-2 items-center justify-between">
-			{#if pagination}
-				<div class="flex items-center gap-2 flex-row">
-					<Button
-						size="xs"
-						variant="border"
-						color="light"
-						on:click={() => $table.previousPage()}
-						disabled={!$table.getCanPreviousPage()}
-					>
-						Previous
-					</Button>
-					<Button
-						size="xs"
-						variant="border"
-						color="light"
-						on:click={() => $table.nextPage()}
-						disabled={!$table.getCanNextPage()}
-					>
-						Next
-					</Button>
-					{$table.getState().pagination.pageIndex + 1} of {$table.getPageCount()}
-				</div>
-			{:else}
-				<div />
-			{/if}
-			<div class="flex items-center gap-2 flex-row">
-				<Button
-					size="xs"
-					variant="border"
-					color="light"
-					on:click={() => {
-						const dataStr =
-							'data:text/json;charset=utf-8,' + encodeURIComponent(JSON.stringify(result))
-						const downloadAnchorNode = document.createElement('a')
-						downloadAnchorNode.setAttribute('href', dataStr)
-						downloadAnchorNode.setAttribute('download', 'data.json')
-						document.body.appendChild(downloadAnchorNode) // required for firefox
-						downloadAnchorNode.click()
-						downloadAnchorNode.remove()
-					}}
-					startIcon={{ icon: faDownload }}
-				>
-					Download
-				</Button>
-			</div>
-		</div>
+
+		<AppTableFooter paginationEnabled={pagination} {result} {table} />
 	</div>
 </RunnableWrapper>
