+++ frontend/src/lib/components/apps/components/table/tableOptions.ts
@@ -1,0 +1,13 @@
+import { getCoreRowModel, getPaginationRowModel } from '@tanstack/svelte-table'
+
+const tableOptions = {
+	getCoreRowModel: getCoreRowModel(),
+	getPaginationRowModel: getPaginationRowModel(),
+	initialState: {
+		pagination: {
+			pageSize: 10
+		}
+	}
+}
+
+export { tableOptions }
