+++ frontend/src/lib/components/apps/editor/AppEditor.svelte
@@ -59,6 +59,7 @@
 	})
 
 	let mounted = false
+
 	onMount(() => {
 		mounted = true
 	})
